ARCHIVE/legacy/logiontology_v2.0.0_initial/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logiontology"
version = "2.0.0"
description = "HVDC Full Stack: Protégé Ontology + Excel→RDF→Neo4j→FastAPI→React + AI Insights + PDF Reports"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [{name="Your Team", email="team@example.com"}]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Typing :: Typed"
]
dependencies = [
  "typer>=0.12.5",
  "pydantic>=2.7.0",
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
  "XlsxWriter>=3.2.0",
  "rdflib>=7.0.0",
  "pyyaml>=6.0.1",
  # Full Stack MVP additions
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "neo4j>=5.14.0",
  "jinja2>=3.1.0",
  "weasyprint>=60.0.0",
  "matplotlib>=3.8.0",
  "httpx>=0.25.0",
  "python-multipart>=0.0.6"
]

[project.optional-dependencies]
shacl = ["pyshacl>=0.23.0"]
api = ["fastapi>=0.104.0", "uvicorn[standard]>=0.24.0"]
graph = ["neo4j>=5.14.0", "rdflib-neo4j>=0.1.0"]
reports = ["jinja2>=3.1.0", "weasyprint>=60.0.0", "matplotlib>=3.8.0"]
ai = ["httpx>=0.25.0", "anthropic>=0.8.0", "openai>=1.0.0"]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.5",
  "black>=24.8.0",
  "isort>=5.13.2",
  "bandit>=1.7.9",
  "pip-audit>=2.7.3",
  "pyupgrade>=3.19.0",
  "mypy>=1.11.0"
]

[project.scripts]
logiontology = "src.cli:app"

[tool.pytest.ini_options]
addopts = "-q --cov=src --cov-report=term-missing"
testpaths = ["tests"]

[tool.isort]
profile = "black"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
target-version = "py313"
line-length = 100
fix = true
exclude = ["docs", ".venv"]

[tool.ruff.format]
docstring-code-format = true

[tool.bandit]
targets = ["src"]
